Refactoring project
