+++ front/manager/index.html
