+++ front/styles.css
@@ -1,0 +1,104 @@
+:root{
+	--background : #283149;
+	--stale: #404b69;
+	--focus: #00818a;
+}
+
+body{
+	font-family: Helvetica;
+	margin: 0;
+	padding: 0;
+	background: var(--background);
+}
+
+form{
+	padding: 150px;
+	width: 500px;
+	max-width: 80%;
+	position: absolute;
+	top: 50%;
+	left: 50%;
+	transform: translate(-50%, -50%);
+	text-align: center;
+}
+
+h1{
+	color: white;
+	font-weight: 500;
+	font-size: 45px;
+}
+
+input{
+	background: none;
+	display: block;
+	margin: 20px auto;
+	border: 2px solid var(--stale);
+	padding: 14px 10px;
+	width: 250px;
+	border-radius: 25px;
+	transition: 0.25s;
+	text-align: center;
+	color: white;
+	font-weight: bold;
+	font-size: 16px;
+}
+input:focus{
+	width: 325px;
+	border-color: var(--focus);
+}
+
+button{
+	background: none;
+	display: block;
+	margin: 20px auto;
+	border: 2px solid var(--focus);
+	padding: 14px 10px;
+	width: 175px;
+	border-radius: 25px;
+	transition: 0.25s;
+	text-align: center;
+	color: white;
+	font-weight: bold;
+	font-size: 16px;
+	cursor: pointer;
+}
+button:hover{
+	background: var(--focus);
+}
+
+.modal{
+	position: fixed;
+	top: 50%;
+	left: 50%;
+	transform: translate(-50%, -50%) scale(0);
+	width: 300px;
+	z-index: 10;
+	border: 3px solid var(--stale);
+	border-radius: 10px;
+	background-color: var(--background);
+	color: white;
+	transition: 0.25s;
+}
+.modal.active{
+	transform: translate(-50%, -50%) scale(1);
+}
+.title{
+	padding: 15px 15px;
+	text-align : center;
+	font-size: 20px;
+}
+#overlay{
+	position: fixed;
+	top: 0;
+	left: 0;
+	bottom: 0;
+	right: 0;
+	background-color: rgba(0,0,0,0.7);
+	opacity: 0;
+	pointer-events: none;
+	transition: 0.25s;
+}
+#overlay.active{
+	opacity: 1;
+	pointer-events: all;
+}
