--- frontend/manager/styles.css
@@ -1,154 +1,0 @@
-:root{
-	--background : #283149;
-	--stale: #404b69;
-	--focus: #00818a;
-}
-
-body{
-	font-family: Helvetica;
-	margin: 0;
-	padding: 0;
-	background: var(--background);
-}
-
-.box{
-	padding: 150px;
-	width: 500px;
-	max-width: 80%;
-	position: absolute;
-	top: 50%;
-	left: 50%;
-	transform: translate(-50%, -50%);
-	text-align: center;
-}
-
-h1{
-	color: white;
-	font-weight: 500;
-	font-size: 45px;
-}
-
-input{
-	background: none;
-	display: block;
-	margin: 20px auto;
-	border: 2px solid var(--stale);
-	padding: 14px 10px;
-	width: 250px;
-	border-radius: 25px;
-	transition: 0.25s;
-	text-align: center;
-	color: white;
-	font-weight: bold;
-	font-size: 16px;
-}
-input:focus{
-	width: 325px;
-	border-color: var(--focus);
-}
-
-button{
-	background: none;
-	display: inline-block;
-	margin: 20px auto;
-	border: 2px solid var(--focus);
-	padding: 14px 10px;
-	width: 175px;
-	border-radius: 25px;
-	transition: 0.25s;
-	text-align: center;
-	color: white;
-	font-weight: bold;
-	font-size: 16px;
-	cursor: pointer;
-}
-button:hover{
-	background: var(--focus);
-}
-
-.modal{
-	position: fixed;
-	top: 50%;
-	left: 50%;
-	transform: translate(-50%, -50%) scale(0);
-	width: 500px;
-	z-index: 10;
-	border: 3px solid var(--stale);
-	border-radius: 10px;
-	background-color: var(--background);
-	color: white;
-	text-align: center;
-	transition: 0.25s;
-	font-size: 20px;
-}
-.modal.active{
-	transform: translate(-50%, -50%) scale(1);
-}
-.modal.error{
-	z-index: 12;
-}
-.title{
-	padding: 15px 15px;
-}
-
-.radio{
-	display: inline-block;
-	vertical-align: middle;
-	position: relative;
-	padding-left: 10px
-}
-.radio + .radio{
-	margin-left: 25px;
-}
-.radio input[type="radio"]{
-	display: none;
-}
-.radio span{
-	width: 16px;
-	height: 16px;
-	border-radius: 50%;
-	border: 3px solid var(--stale);
-	display: block;
-	position: absolute;
-	left: -15px;
-	top: 0;
-}
-.radio span:after{
-	content: "";
-	height: 10px;
-	width: 10px;
-	background-color: var(--focus);
-	display: block;
-	position: absolute;
-	left: 50%;
-	top: 50%;
-	transform: translate(-50%, -50%) scale(0);
-	border-radius: 50%;
-	transition: 0.25s ease-in-out 0s;
-}
-.radio input[type="radio"]:checked ~ span:after{
-	transform: translate(-50%, -50%) scale(1);
-}
-
-
-
-
-#overlay{
-	position: fixed;
-	top: 0;
-	left: 0;
-	bottom: 0;
-	right: 0;
-	background-color: rgba(0,0,0,0.7);
-	opacity: 0;
-	pointer-events: none;
-	transition: 0.25s;
-	z-index: 9;
-}
-#overlay.active{
-	opacity: 1;
-	pointer-events: all;
-}
-#overlay.error{
-	z-index: 11;
-}
